Rename packages for EPICS serialisation tests
